pyproject.toml
[project]
name = "midjourney-bridge"
version = "0.1.0"
description = "Open-source Python client + Claude skill + MCP server that lets your AI tooling drive your own Midjourney subscription via the midjourney.com web API."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Like A Human", email = "academy@likeahuman.ai" }]
keywords = ["midjourney", "mcp", "claude", "ai", "image-generation"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "curl-cffi>=0.13.0",
    "pydantic>=2.7",
    "platformdirs>=4.2",
    "python-dotenv>=1.0",
    "rapidfuzz>=3.9",
    "typer>=0.12",
    "mcp>=1.2",
    "browser-cookie3>=0.20",
]

[project.urls]
Homepage = "https://github.com/likeahuman-ai/midjourney-bridge"
Repository = "https://github.com/likeahuman-ai/midjourney-bridge"
Issues = "https://github.com/likeahuman-ai/midjourney-bridge/issues"

[project.scripts]
mj = "midjourney_bridge.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/midjourney_bridge"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "mypy>=1.10",
    "ruff>=0.5",
    "typer>=0.25.1",
    "mcp>=1.27.0",
    "pytest-asyncio>=1.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["curl_cffi.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["browser_cookie3.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra --strict-markers -m 'not integration'"
asyncio_mode = "auto"
markers = [
    "integration: hits real Midjourney API (requires MJ_COOKIE; skipped by default)",
]
